+++ YASAM.SteamInterface/Models/Game.cs
@@ -1,3 +1,3 @@
 namespace YASAM.SteamInterface;
 
-public record Game(uint AppId, string Name, bool Installed, string ImgUrl, int PlaytimeTotal);
+public record Game(ulong AppId, string Name, bool Installed, string ImgUrl, int PlaytimeTotal);
